Move swizzle handling to utility class
